Docs: Homebrew setup: do not add brew commands to the shell profile
The recommended Homebrew setup required running two `brew --prefix` commands
to the shell profile. `brew` commands are slow, so this introduced a notable
lag when sourcing the shell profile.

Homebrew paths do not change, so we can determine the necessary paths once
during setup and use static paths in the shell profile.